rename log -> tail add makefile
